refactor: auto-split test_third_party_integration_audit.py
