+++ tests/unit/third_party_integration_audit/test_cross_engine_protocol_compliance.py
@@ -1,0 +1,147 @@
+"""Comprehensive third-party integration audit tests.
+
+This module provides adversarial TDD tests for all third-party package
+integrations in UpstreamDrift. Tests are organized by package and
+structured to verify:
+
+1. Import resilience (graceful degradation when packages missing)
+2. Protocol compliance (PhysicsEngine interface adherence)
+3. API correctness (correct method signatures and return types)
+4. Error handling (proper exceptions, no silent failures)
+
+Issues: #1810, #1811, #1812, #1813, #1814, #1815, #1816, #1817, #1818
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.engine_core.engine_availability import (
+    DRAKE_AVAILABLE,
+    MEDIAPIPE_AVAILABLE,
+    MUJOCO_AVAILABLE,
+    MYOSUITE_AVAILABLE,
+    OPENSIM_AVAILABLE,
+    PINOCCHIO_AVAILABLE,
+    get_available_engines,
+    get_unavailable_engines,
+    is_engine_available,
+    skip_if_unavailable,
+)
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 1. Engine Availability Infrastructure Tests (#1818)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 2. Drake Integration Tests (#1810)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 3. MuJoCo Integration Tests (#1811)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 4. Pinocchio Integration Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 5. Pink IK Solver Tests (#1812 sub-component)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 6. OpenSim Integration Tests (#1813)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 7. MyoSuite Integration Tests (#1814)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 8. OpenPose Integration Tests (#1815)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 9. MediaPipe Integration Tests (#1816)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 10. Pose Estimation Interface Tests (#1817)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 11. Cross-Engine Protocol Compliance Tests
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+class TestCrossEngineProtocolCompliance:
+    """Verify all engines adhere to the PhysicsEngine protocol."""
+
+    PHYSICS_ENGINE_METHODS = [
+        "load_from_path",
+        "load_from_string",
+        "reset",
+        "step",
+        "forward",
+        "get_state",
+        "set_state",
+        "set_control",
+        "get_time",
+        "compute_mass_matrix",
+        "compute_bias_forces",
+        "compute_inverse_dynamics",
+        "compute_jacobian",
+    ]
+
+    @pytest.mark.parametrize(
+        "engine_module,engine_class",
+        [
+            (
+                "src.engines.physics_engines.drake.python.drake_physics_engine",
+                "DrakePhysicsEngine",
+            ),
+            (
+                "src.engines.physics_engines.pinocchio.python.pinocchio_physics_engine",
+                "PinocchioPhysicsEngine",
+            ),
+            (
+                "src.engines.physics_engines.opensim.python.opensim_physics_engine",
+                "OpenSimPhysicsEngine",
+            ),
+            (
+                "src.engines.physics_engines.myosuite.python.myosuite_physics_engine",
+                "MyoSuitePhysicsEngine",
+            ),
+        ],
+    )
+    def test_engine_has_all_protocol_methods(
+        self, engine_module: str, engine_class: str
+    ) -> None:
+        """Every engine must implement all PhysicsEngine protocol methods."""
+        import importlib
+
+        mod = importlib.import_module(engine_module)
+        cls = getattr(mod, engine_class)
+
+        for method in self.PHYSICS_ENGINE_METHODS:
+            assert hasattr(cls, method), (
+                f"{engine_class} missing protocol method: {method}"
+            )
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 12. dtack Subpackage Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
