+++ tests/unit/third_party_integration_audit/test_dtack_subpackage_audit.py
@@ -1,0 +1,142 @@
+"""Comprehensive third-party integration audit tests.
+
+This module provides adversarial TDD tests for all third-party package
+integrations in UpstreamDrift. Tests are organized by package and
+structured to verify:
+
+1. Import resilience (graceful degradation when packages missing)
+2. Protocol compliance (PhysicsEngine interface adherence)
+3. API correctness (correct method signatures and return types)
+4. Error handling (proper exceptions, no silent failures)
+
+Issues: #1810, #1811, #1812, #1813, #1814, #1815, #1816, #1817, #1818
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.engine_core.engine_availability import (
+    DRAKE_AVAILABLE,
+    MEDIAPIPE_AVAILABLE,
+    MUJOCO_AVAILABLE,
+    MYOSUITE_AVAILABLE,
+    OPENSIM_AVAILABLE,
+    PINOCCHIO_AVAILABLE,
+    get_available_engines,
+    get_unavailable_engines,
+    is_engine_available,
+    skip_if_unavailable,
+)
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 1. Engine Availability Infrastructure Tests (#1818)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 2. Drake Integration Tests (#1810)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 3. MuJoCo Integration Tests (#1811)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 4. Pinocchio Integration Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 5. Pink IK Solver Tests (#1812 sub-component)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 6. OpenSim Integration Tests (#1813)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 7. MyoSuite Integration Tests (#1814)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 8. OpenPose Integration Tests (#1815)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 9. MediaPipe Integration Tests (#1816)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 10. Pose Estimation Interface Tests (#1817)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 11. Cross-Engine Protocol Compliance Tests
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 12. dtack Subpackage Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+class TestDtackSubpackageAudit:
+    """Verify pinocchio dtack subpackage integrity."""
+
+    def test_dtack_init_importable(self) -> None:
+        """dtack __init__.py must be importable."""
+        from src.engines.physics_engines.pinocchio.python import dtack
+
+        assert dtack is not None
+
+    def test_dtack_backends_importable(self) -> None:
+        """dtack backends package must be importable."""
+        from src.engines.physics_engines.pinocchio.python.dtack import backends
+
+        assert backends is not None
+
+    def test_dtack_mujoco_backend_has_import_guard(self) -> None:
+        """MuJoCoBackend must not crash when mujoco is not installed."""
+        from src.engines.physics_engines.pinocchio.python.dtack.backends.mujoco_backend import (  # noqa: E501
+            MuJoCoBackend,
+        )
+
+        assert MuJoCoBackend is not None
+
+    def test_dtack_backend_factory_importable(self) -> None:
+        """BackendFactory must be importable."""
+        from src.engines.physics_engines.pinocchio.python.dtack.backends.backend_factory import (  # noqa: E501
+            BackendFactory,
+            BackendType,
+        )
+
+        assert BackendFactory is not None
+        assert BackendType is not None
+        # Verify enum values
+        assert BackendType.PINOCCHIO == "pinocchio"
+        assert BackendType.MUJOCO == "mujoco"
+        assert BackendType.PINK == "pink"
+
+    def test_dtack_ik_importable(self) -> None:
+        """dtack ik package must be importable."""
+        from src.engines.physics_engines.pinocchio.python.dtack import ik
+
+        assert ik is not None
+
+    def test_dtack_utils_importable(self) -> None:
+        """dtack utils package must be importable."""
+        from src.engines.physics_engines.pinocchio.python.dtack import utils
+
+        assert utils is not None
