+++ tests/unit/third_party_integration_audit/test_engine_availability_infrastructure.py
@@ -1,0 +1,165 @@
+"""Comprehensive third-party integration audit tests.
+
+This module provides adversarial TDD tests for all third-party package
+integrations in UpstreamDrift. Tests are organized by package and
+structured to verify:
+
+1. Import resilience (graceful degradation when packages missing)
+2. Protocol compliance (PhysicsEngine interface adherence)
+3. API correctness (correct method signatures and return types)
+4. Error handling (proper exceptions, no silent failures)
+
+Issues: #1810, #1811, #1812, #1813, #1814, #1815, #1816, #1817, #1818
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.engine_core.engine_availability import (
+    DRAKE_AVAILABLE,
+    MEDIAPIPE_AVAILABLE,
+    MUJOCO_AVAILABLE,
+    MYOSUITE_AVAILABLE,
+    OPENSIM_AVAILABLE,
+    PINOCCHIO_AVAILABLE,
+    get_available_engines,
+    get_unavailable_engines,
+    is_engine_available,
+    skip_if_unavailable,
+)
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 1. Engine Availability Infrastructure Tests (#1818)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+class TestEngineAvailabilityInfrastructure:
+    """Verify engine_availability.py correctness and consistency."""
+
+    def test_is_engine_available_case_insensitive(self) -> None:
+        """Engine name lookup must be case-insensitive."""
+        # numpy is almost always available
+        result_lower = is_engine_available("numpy")
+        result_upper = is_engine_available("NUMPY")
+        result_mixed = is_engine_available("NumPy")
+        assert result_lower == result_upper == result_mixed
+
+    def test_get_available_engines_returns_list(self) -> None:
+        """get_available_engines must return a list, never None."""
+        result = get_available_engines()
+        assert isinstance(result, list)
+        # numpy should always be available
+        assert "numpy" in result
+
+    def test_get_unavailable_engines_returns_list(self) -> None:
+        """get_unavailable_engines must return a list."""
+        result = get_unavailable_engines()
+        assert isinstance(result, list)
+
+    def test_available_and_unavailable_are_disjoint(self) -> None:
+        """No engine should appear in both available and unavailable lists."""
+        available = set(get_available_engines())
+        unavailable = set(get_unavailable_engines())
+        overlap = available & unavailable
+        assert not overlap, f"Engines in both lists: {overlap}"
+
+    def test_openpose_availability_check_uses_pyopenpose(self) -> None:
+        """OpenPose availability check must use 'pyopenpose', not 'openpose'.
+
+        The Python bindings for OpenPose are distributed under the module
+        name 'pyopenpose', not 'openpose'. Using the wrong name would
+        always report unavailable even when installed.
+        """
+        # Verify the import in engine_availability.py uses pyopenpose
+        import inspect
+
+        from src.shared.python.engine_core import engine_availability
+
+        source = inspect.getsource(engine_availability)
+        assert "import pyopenpose" in source, (
+            "engine_availability.py should check for 'pyopenpose', not 'openpose'"
+        )
+
+    def test_skip_if_unavailable_returns_marker(self) -> None:
+        """skip_if_unavailable must return a pytest marker."""
+        marker = skip_if_unavailable("numpy")
+        assert hasattr(marker, "mark") or hasattr(marker, "args")
+
+    def test_mediapipe_flag_exists(self) -> None:
+        """MEDIAPIPE_AVAILABLE flag must exist and be boolean."""
+        assert isinstance(MEDIAPIPE_AVAILABLE, bool)
+
+    def test_engine_flags_dict_has_all_physics_engines(self) -> None:
+        """_ENGINE_FLAGS must include all core physics engines."""
+        from src.shared.python.engine_core.engine_availability import _ENGINE_FLAGS
+
+        required_engines = [
+            "mujoco",
+            "pinocchio",
+            "drake",
+            "opensim",
+            "myosuite",
+            "mediapipe",
+            "openpose",
+        ]
+        for engine in required_engines:
+            assert engine in _ENGINE_FLAGS, f"Missing engine flag: {engine}"
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 2. Drake Integration Tests (#1810)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 3. MuJoCo Integration Tests (#1811)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 4. Pinocchio Integration Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 5. Pink IK Solver Tests (#1812 sub-component)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 6. OpenSim Integration Tests (#1813)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 7. MyoSuite Integration Tests (#1814)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 8. OpenPose Integration Tests (#1815)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 9. MediaPipe Integration Tests (#1816)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 10. Pose Estimation Interface Tests (#1817)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 11. Cross-Engine Protocol Compliance Tests
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 12. dtack Subpackage Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
