+++ tests/unit/third_party_integration_audit/test_media_pipe_integration_audit.py
@@ -1,0 +1,182 @@
+"""Comprehensive third-party integration audit tests.
+
+This module provides adversarial TDD tests for all third-party package
+integrations in UpstreamDrift. Tests are organized by package and
+structured to verify:
+
+1. Import resilience (graceful degradation when packages missing)
+2. Protocol compliance (PhysicsEngine interface adherence)
+3. API correctness (correct method signatures and return types)
+4. Error handling (proper exceptions, no silent failures)
+
+Issues: #1810, #1811, #1812, #1813, #1814, #1815, #1816, #1817, #1818
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.engine_core.engine_availability import (
+    DRAKE_AVAILABLE,
+    MEDIAPIPE_AVAILABLE,
+    MUJOCO_AVAILABLE,
+    MYOSUITE_AVAILABLE,
+    OPENSIM_AVAILABLE,
+    PINOCCHIO_AVAILABLE,
+    get_available_engines,
+    get_unavailable_engines,
+    is_engine_available,
+    skip_if_unavailable,
+)
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 1. Engine Availability Infrastructure Tests (#1818)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 2. Drake Integration Tests (#1810)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 3. MuJoCo Integration Tests (#1811)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 4. Pinocchio Integration Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 5. Pink IK Solver Tests (#1812 sub-component)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 6. OpenSim Integration Tests (#1813)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 7. MyoSuite Integration Tests (#1814)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 8. OpenPose Integration Tests (#1815)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 9. MediaPipe Integration Tests (#1816)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+class TestMediaPipeIntegrationAudit:
+    """Verify MediaPipe integration correctness."""
+
+    def test_mediapipe_estimator_importable(self) -> None:
+        """MediaPipeEstimator must be importable."""
+        from src.shared.python.pose_estimation.mediapipe_estimator import (
+            MediaPipeEstimator,
+        )
+
+        assert MediaPipeEstimator is not None
+
+    def test_mediapipe_estimator_instantiation(self) -> None:
+        """MediaPipeEstimator must instantiate with default parameters."""
+        from src.shared.python.pose_estimation.mediapipe_estimator import (
+            MediaPipeEstimator,
+        )
+
+        estimator = MediaPipeEstimator()
+        assert estimator is not None
+        assert estimator.enable_temporal_smoothing is True  # noqa: E712
+
+    def test_mediapipe_estimator_custom_params(self) -> None:
+        """MediaPipeEstimator must accept custom parameters."""
+        from src.shared.python.pose_estimation.mediapipe_estimator import (
+            MediaPipeEstimator,
+        )
+
+        estimator = MediaPipeEstimator(
+            min_detection_confidence=0.7,
+            min_tracking_confidence=0.8,
+            enable_temporal_smoothing=False,
+        )
+        assert estimator.enable_temporal_smoothing is False  # noqa: E712
+
+    def test_mediapipe_reset_temporal_state_exists(self) -> None:
+        """reset_temporal_state must exist and be callable."""
+        from src.shared.python.pose_estimation.mediapipe_estimator import (
+            MediaPipeEstimator,
+        )
+
+        estimator = MediaPipeEstimator()
+        assert hasattr(estimator, "reset_temporal_state")
+        assert callable(estimator.reset_temporal_state)
+        # Should not crash when called
+        estimator.reset_temporal_state()
+
+    def test_mediapipe_implements_pose_estimator_interface(self) -> None:
+        """MediaPipeEstimator must implement PoseEstimator ABC."""
+        from src.shared.python.pose_estimation.interface import PoseEstimator
+        from src.shared.python.pose_estimation.mediapipe_estimator import (
+            MediaPipeEstimator,
+        )
+
+        assert issubclass(MediaPipeEstimator, PoseEstimator)
+
+    def test_mediapipe_video_resets_kalman_state(self) -> None:
+        """estimate_from_video must call reset_temporal_state at start.
+
+        This prevents Kalman filter contamination between video files.
+        """
+        import inspect
+
+        from src.shared.python.pose_estimation.mediapipe_estimator import (
+            MediaPipeEstimator,
+        )
+
+        source = inspect.getsource(MediaPipeEstimator.estimate_from_video)
+        assert "reset_temporal_state" in source, (
+            "estimate_from_video must call reset_temporal_state() at start "
+            "to prevent Kalman filter contamination between videos"
+        )
+
+    def test_mediapipe_reset_clears_kalman_filters(self) -> None:
+        """reset_temporal_state must clear all Kalman filter state."""
+        from src.shared.python.pose_estimation.mediapipe_estimator import (
+            MediaPipeEstimator,
+        )
+
+        estimator = MediaPipeEstimator()
+        # Simulate having some state
+        estimator.kalman_filters["test"] = MagicMock()
+        estimator.previous_landmarks = {"test": np.array([1, 2, 3])}
+
+        # Reset
+        estimator.reset_temporal_state()
+
+        assert len(estimator.kalman_filters) == 0
+        assert estimator.previous_landmarks is None
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 10. Pose Estimation Interface Tests (#1817)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 11. Cross-Engine Protocol Compliance Tests
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 12. dtack Subpackage Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
