+++ tests/unit/third_party_integration_audit/test_mu_jo_co_integration_audit.py
@@ -1,0 +1,121 @@
+"""Comprehensive third-party integration audit tests.
+
+This module provides adversarial TDD tests for all third-party package
+integrations in UpstreamDrift. Tests are organized by package and
+structured to verify:
+
+1. Import resilience (graceful degradation when packages missing)
+2. Protocol compliance (PhysicsEngine interface adherence)
+3. API correctness (correct method signatures and return types)
+4. Error handling (proper exceptions, no silent failures)
+
+Issues: #1810, #1811, #1812, #1813, #1814, #1815, #1816, #1817, #1818
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.engine_core.engine_availability import (
+    DRAKE_AVAILABLE,
+    MEDIAPIPE_AVAILABLE,
+    MUJOCO_AVAILABLE,
+    MYOSUITE_AVAILABLE,
+    OPENSIM_AVAILABLE,
+    PINOCCHIO_AVAILABLE,
+    get_available_engines,
+    get_unavailable_engines,
+    is_engine_available,
+    skip_if_unavailable,
+)
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 1. Engine Availability Infrastructure Tests (#1818)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 2. Drake Integration Tests (#1810)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 3. MuJoCo Integration Tests (#1811)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+class TestMuJoCoIntegrationAudit:
+    """Verify MuJoCo integration correctness."""
+
+    def test_mujoco_availability_flag_is_boolean(self) -> None:
+        """MUJOCO_AVAILABLE must be a boolean."""
+        assert isinstance(MUJOCO_AVAILABLE, bool)
+
+    def test_mujoco_conftest_early_import(self) -> None:
+        """conftest.py must attempt early MuJoCo import to avoid DLL conflicts."""
+        conftest_path = Path(__file__).parent.parent.parent / "conftest.py"
+        if conftest_path.exists():
+            content = conftest_path.read_text(encoding="utf-8")
+            assert "mujoco" in content, (
+                "conftest.py should import mujoco early to avoid DLL conflicts"
+            )
+
+    @skip_if_unavailable("mujoco")
+    def test_mujoco_gl_environment(self) -> None:
+        """MUJOCO_GL should not be set to 'egl' on Windows (headless)."""
+        import os
+        import sys
+
+        if sys.platform == "win32":
+            gl = os.environ.get("MUJOCO_GL", "")
+            assert gl != "egl", (
+                "MUJOCO_GL=egl is invalid on Windows; use osmesa or glfw"
+            )
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 4. Pinocchio Integration Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 5. Pink IK Solver Tests (#1812 sub-component)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 6. OpenSim Integration Tests (#1813)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 7. MyoSuite Integration Tests (#1814)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 8. OpenPose Integration Tests (#1815)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 9. MediaPipe Integration Tests (#1816)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 10. Pose Estimation Interface Tests (#1817)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 11. Cross-Engine Protocol Compliance Tests
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 12. dtack Subpackage Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
