+++ tests/unit/third_party_integration_audit/test_open_sim_integration_audit.py
@@ -1,0 +1,153 @@
+"""Comprehensive third-party integration audit tests.
+
+This module provides adversarial TDD tests for all third-party package
+integrations in UpstreamDrift. Tests are organized by package and
+structured to verify:
+
+1. Import resilience (graceful degradation when packages missing)
+2. Protocol compliance (PhysicsEngine interface adherence)
+3. API correctness (correct method signatures and return types)
+4. Error handling (proper exceptions, no silent failures)
+
+Issues: #1810, #1811, #1812, #1813, #1814, #1815, #1816, #1817, #1818
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.engine_core.engine_availability import (
+    DRAKE_AVAILABLE,
+    MEDIAPIPE_AVAILABLE,
+    MUJOCO_AVAILABLE,
+    MYOSUITE_AVAILABLE,
+    OPENSIM_AVAILABLE,
+    PINOCCHIO_AVAILABLE,
+    get_available_engines,
+    get_unavailable_engines,
+    is_engine_available,
+    skip_if_unavailable,
+)
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 1. Engine Availability Infrastructure Tests (#1818)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 2. Drake Integration Tests (#1810)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 3. MuJoCo Integration Tests (#1811)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 4. Pinocchio Integration Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 5. Pink IK Solver Tests (#1812 sub-component)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 6. OpenSim Integration Tests (#1813)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+class TestOpenSimIntegrationAudit:
+    """Verify OpenSim integration correctness."""
+
+    def test_opensim_availability_flag_is_boolean(self) -> None:
+        """OPENSIM_AVAILABLE must be a boolean."""
+        assert isinstance(OPENSIM_AVAILABLE, bool)
+
+    def test_opensim_engine_importable(self) -> None:
+        """OpenSimPhysicsEngine must be importable."""
+        from src.engines.physics_engines.opensim.python.opensim_physics_engine import (
+            OpenSimPhysicsEngine,
+        )
+
+        assert OpenSimPhysicsEngine is not None
+
+    @skip_if_unavailable("opensim")
+    def test_opensim_engine_initialization(self) -> None:
+        """OpenSimPhysicsEngine must initialize without model."""
+        from src.engines.physics_engines.opensim.python.opensim_physics_engine import (
+            OpenSimPhysicsEngine,
+        )
+
+        engine = OpenSimPhysicsEngine()
+        assert not engine.is_initialized
+        assert engine.model_name == ""
+
+    @skip_if_unavailable("opensim")
+    def test_opensim_protocol_methods_exist(self) -> None:
+        """OpenSimPhysicsEngine must implement all PhysicsEngine protocol methods."""
+        from src.engines.physics_engines.opensim.python.opensim_physics_engine import (
+            OpenSimPhysicsEngine,
+        )
+
+        engine = OpenSimPhysicsEngine()
+        required_methods = [
+            "load_from_path",
+            "load_from_string",
+            "reset",
+            "step",
+            "forward",
+            "get_state",
+            "set_state",
+            "set_control",
+            "get_time",
+            "compute_mass_matrix",
+            "compute_bias_forces",
+            "compute_gravity_forces",
+            "compute_inverse_dynamics",
+            "compute_jacobian",
+            "compute_drift_acceleration",
+        ]
+        for method in required_methods:
+            assert hasattr(engine, method), f"Missing method: {method}"
+
+    def test_opensim_muscle_analysis_importable(self) -> None:
+        """muscle_analysis.py must be importable."""
+        from src.engines.physics_engines.opensim.python import muscle_analysis
+
+        assert muscle_analysis is not None
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 7. MyoSuite Integration Tests (#1814)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 8. OpenPose Integration Tests (#1815)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 9. MediaPipe Integration Tests (#1816)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 10. Pose Estimation Interface Tests (#1817)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 11. Cross-Engine Protocol Compliance Tests
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 12. dtack Subpackage Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
