+++ tests/unit/third_party_integration_audit/test_pose_estimation_interface_audit.py
@@ -1,0 +1,144 @@
+"""Comprehensive third-party integration audit tests.
+
+This module provides adversarial TDD tests for all third-party package
+integrations in UpstreamDrift. Tests are organized by package and
+structured to verify:
+
+1. Import resilience (graceful degradation when packages missing)
+2. Protocol compliance (PhysicsEngine interface adherence)
+3. API correctness (correct method signatures and return types)
+4. Error handling (proper exceptions, no silent failures)
+
+Issues: #1810, #1811, #1812, #1813, #1814, #1815, #1816, #1817, #1818
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.engine_core.engine_availability import (
+    DRAKE_AVAILABLE,
+    MEDIAPIPE_AVAILABLE,
+    MUJOCO_AVAILABLE,
+    MYOSUITE_AVAILABLE,
+    OPENSIM_AVAILABLE,
+    PINOCCHIO_AVAILABLE,
+    get_available_engines,
+    get_unavailable_engines,
+    is_engine_available,
+    skip_if_unavailable,
+)
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 1. Engine Availability Infrastructure Tests (#1818)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 2. Drake Integration Tests (#1810)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 3. MuJoCo Integration Tests (#1811)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 4. Pinocchio Integration Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 5. Pink IK Solver Tests (#1812 sub-component)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 6. OpenSim Integration Tests (#1813)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 7. MyoSuite Integration Tests (#1814)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 8. OpenPose Integration Tests (#1815)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 9. MediaPipe Integration Tests (#1816)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 10. Pose Estimation Interface Tests (#1817)
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+class TestPoseEstimationInterfaceAudit:
+    """Verify pose estimation interface and shared utilities."""
+
+    def test_interface_importable(self) -> None:
+        """PoseEstimator and PoseEstimationResult must be importable."""
+        from src.shared.python.pose_estimation.interface import (
+            PoseEstimationResult,
+            PoseEstimator,
+        )
+
+        assert PoseEstimator is not None
+        assert PoseEstimationResult is not None
+
+    def test_pose_estimation_result_fields(self) -> None:
+        """PoseEstimationResult must have required fields."""
+        from src.shared.python.pose_estimation.interface import PoseEstimationResult
+
+        result = PoseEstimationResult(
+            joint_angles={"elbow": 1.5},
+            confidence=0.9,
+            timestamp=1.0,
+            raw_keypoints=None,
+        )
+        assert result.joint_angles == {"elbow": 1.5}
+        assert result.confidence == pytest.approx(0.9)
+        assert result.timestamp == pytest.approx(1.0)
+        assert result.raw_keypoints is None
+
+    def test_joint_angle_utils_importable(self) -> None:
+        """Joint angle utilities must be importable."""
+        from src.shared.python.pose_estimation.joint_angle_utils import (
+            compute_joint_angles,
+        )
+
+        assert compute_joint_angles is not None
+
+    def test_openpose_canonical_mapping_exists(self) -> None:
+        """OPENPOSE_TO_CANONICAL mapping must exist."""
+        from src.shared.python.pose_estimation.joint_angle_utils import (
+            OPENPOSE_TO_CANONICAL,
+        )
+
+        assert isinstance(OPENPOSE_TO_CANONICAL, dict)
+        assert len(OPENPOSE_TO_CANONICAL) > 0
+
+    def test_validation_metrics_importable(self) -> None:
+        """Validation metrics module must be importable."""
+        from src.shared.python.pose_estimation import validation_metrics
+
+        assert validation_metrics is not None
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 11. Cross-Engine Protocol Compliance Tests
+# ═══════════════════════════════════════════════════════════════════════════════
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# 12. dtack Subpackage Tests (#1812)
+# ═══════════════════════════════════════════════════════════════════════════════
